lib: Introduce console device-backed file library
Introduce a new `lib/ukfile-console` library that provides
console device-backed `uk_file` instances. The serial file
implementation previously residing in `lib/posix-tty/serial.c`
is moved here so that any subsystem — not just posix-tty — can
obtain a file abstraction over a registered `uk_console` device.

The static `uk_file_serial` is replaced with a dynamically
allocated console file, created via the new public API:

  const struct uk_file *uk_file_console_create(struct uk_console *cons);

Each call allocates a `consolefd_alloc` object containing the
file state, refcount, and a back-pointer to the `uk_console`
device. The console pointer is stored in `f->node` and used
directly by the read/write/ctl ops, removing the need for a
global `tty_cons` variable.

The CR injection and echo-back config options are migrated from
`LIBPOSIX_TTY_OUT_NOCARRIAGERETURN`/`LIBPOSIX_TTY_IN_NOECHOBACK`
to `LIBUKFILE_CONSOLE_OUT_NOCARRIAGERETURN`/`LIBUKFILE_CONSOLE_IN_NOECHOBACK`.

`lib/posix-tty` is updated to use `uk_file_console_create()` and
the `tty_f` global file pointer, and `init_tty_cons` is moved to
a `uk_rootfs_initcall` at `UK_FS_PRIO_FSAVAIL` priority.

Checkpatch-Ignore: COMMIT_LOG_LONG_LINE
Checkpatch-Ignore: AVOID_EXTERNS
Checkpatch-Ignore: ENOSYS
