+++ lib/Makefile.uk
@@ -89,3 +89,4 @@ $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukpaging))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/uklcpu))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukpm))
 $(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukpcpuvar))
+$(eval $(call import_lib,$(CONFIG_UK_BASE)/lib/ukfile-console))
