+++ lib/posix-tty/Config.uk
@@ -15,6 +15,7 @@ config LIBPOSIX_TTY_DEVFS
 config LIBPOSIX_TTY_SERIAL
 	bool "Provide file driver for the serial console"
 	select LIBUKCONSOLE
+	select LIBUKFILE_CONSOLE
 
 config LIBPOSIX_TTY_STDIO
 	bool "Open standard I/O files at boot (fd 0, 1, 2)"
@@ -58,11 +59,5 @@ config LIBPOSIX_TTY_STDOUT_NULL
 endchoice
 
 endif
-
-config LIBPOSIX_TTY_OUT_NOCARRIAGERETURN
-	bool "Serial output: no carriage return"
-
-config LIBPOSIX_TTY_IN_NOECHOBACK
-	bool "Serial input: no echo back"
 
 endif
