+++ lib/posix-tty/tty.c
@@ -8,7 +8,6 @@
 #include <uk/init.h>
 #include <uk/posix-fd.h>
 #include <uk/posix-fdtab.h>
-#include <uk/posix-serialfile.h>
 
 #if CONFIG_LIBPOSIX_TTY_PSEUDO
 #include <uk/file-pseudo.h>
@@ -26,9 +25,9 @@
 #define STDERR_DEVNAME "stderr"
 #define STDERR_DEVNAME_LEN (sizeof(STDERR_DEVNAME) - 1)
 
-static int init_posix_tty_devfs(const struct uk_file *in,
-				const struct uk_file *out,
-				const struct uk_file *err)
+static int init_posix_tty_devfs_stdio(const struct uk_file *in,
+				      const struct uk_file *out,
+				      const struct uk_file *err)
 {
 	const void *r;
 
@@ -78,6 +77,8 @@ static int init_posix_tty_devfs(const struct uk_file *in,
 #define STDOUT_FNAME_SERIAL "stdout:serial"
 #define STDOUT_FNAME_LEN_SERIAL (sizeof(STDOUT_FNAME_SERIAL) - 1)
 
+extern const struct uk_file *tty_f;
+
 static int init_posix_tty(struct uk_init_ctx *ictx __unused)
 {
 	const struct uk_file *in;
@@ -97,7 +98,7 @@ static int init_posix_tty(struct uk_init_ctx *ictx __unused)
 	in_fname = STDIN_FNAME_VOID;
 	in_fnamelen = STDIN_FNAME_LEN_VOID;
 #elif CONFIG_LIBPOSIX_TTY_STDIN_SERIAL
-	in = &uk_file_serial;
+	in = tty_f;
 	in_fname = STDIN_FNAME_SERIAL;
 	in_fnamelen = STDIN_FNAME_LEN_SERIAL;
 #else /* !CONFIG_LIBPOSIX_TTY_STDIN_* */
@@ -109,7 +110,7 @@ static int init_posix_tty(struct uk_init_ctx *ictx __unused)
 	out_fname = STDOUT_FNAME_NULL;
 	out_fnamelen = STDOUT_FNAME_LEN_NULL;
 #elif CONFIG_LIBPOSIX_TTY_STDOUT_SERIAL
-	out = &uk_file_serial;
+	out = tty_f;
 	out_fname = STDOUT_FNAME_SERIAL;
 	out_fnamelen = STDOUT_FNAME_LEN_SERIAL;
 #else /* !CONFIG_LIBPOSIX_TTY_STDOUT_* */
@@ -137,14 +138,18 @@ static int init_posix_tty(struct uk_init_ctx *ictx __unused)
 	}
 
 #if CONFIG_LIBPOSIX_TTY_DEVFS
-	return init_posix_tty_devfs(in, out, out);
+	return init_posix_tty_devfs_stdio(in, out, out);
 #else /* !CONFIG_LIBPOSIX_TTY_DEVFS */
 	return 0;
 #endif /* !CONFIG_LIBPOSIX_TTY_DEVFS */
 }
 
+#if CONFIG_LIBPOSIX_TTY_SERIAL
+uk_rootfs_initcall_prio(init_posix_tty, 0x0, UK_PRIO_AFTER(UK_FS_PRIO_FSAVAIL));
+#else /* !CONFIG_LIBPOSIX_TTY_SERIAL */
 #if CONFIG_LIBPOSIX_TTY_DEVFS
 uk_rootfs_initcall_prio(init_posix_tty, 0x0, UK_FS_PRIO_FSAVAIL);
 #else /* !CONFIG_LIBPOSIX_TTY_DEVFS */
 uk_rootfs_initcall(init_posix_tty, 0x0);
 #endif /* !CONFIG_LIBPOSIX_TTY_DEVFS */
+#endif /* !CONFIG_LIBPOSIX_TTY_SERIAL */
