+++ lib/ukfile-console/Makefile.uk
@@ -1,0 +1,6 @@
+$(eval $(call addlib_s,libukfile_console,$(CONFIG_LIBUKFILE_CONSOLE)))
+
+CINCLUDES-$(CONFIG_LIBUKFILE_CONSOLE) += -I$(LIBUKFILE_CONSOLE_BASE)/include
+CXXINCLUDES-$(CONFIG_LIBUKFILE_CONSOLE) += -I$(LIBUKFILE_CONSOLE_BASE)/include
+
+LIBUKFILE_CONSOLE_SRCS-y += $(LIBUKFILE_CONSOLE_BASE)/console.c
