+++ lib/ukfile-console/exportsyms.uk
@@ -1,0 +1,1 @@
+uk_file_console_create
